Move dashboard LearningObject to new dashboard
